+++ pulsar/utils/httpurl.py
@@ -31,10 +31,6 @@ HTTP Parser
    :members:
    :member-order: bysource
 
-parse cookie
-~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: parse_cookie
 
 .. _http-parser: https://github.com/benoitc/http-parser
 .. _urllib3: https://github.com/shazow/urllib3
@@ -61,13 +57,17 @@ from .structures import mapping_iterator, OrderedDict
 from .pep import ispy3k, iteritems, itervalues, to_bytes, native_str
 from .html import capfirst
 
-try:
-    from http_parser.parser import HttpParser as CHttpParser
-    hasextensions = True
-    _Http_Parser = CHttpParser
-except ImportError:  # pragma    nocover
-    hasextensions = False
-    _Http_Parser = None
+#try:
+#    from http_parser.parser import HttpParser as CHttpParser
+#    hasextensions = True
+#    _Http_Parser = CHttpParser
+#except ImportError:  # pragma    nocover
+#    hasextensions = False
+#    _Http_Parser = None
+#
+#The http_parser has several bus, therefore it is switched off
+hasextensions = False
+_Http_Parser = None
 
 try:
     from select import poll, POLLIN
@@ -107,7 +107,7 @@ if ispy3k:  # Python 3
                               urljoin)
     from http.client import responses
     from http.cookiejar import CookieJar, Cookie
-    from http.cookies import SimpleCookie, BaseCookie, CookieError
+    from http.cookies import SimpleCookie
 
     string_type = str
     getproxies_environment = urllibr.getproxies_environment
@@ -132,7 +132,7 @@ else:   # pragma : no cover
                           parse_qsl)
     from httplib import responses
     from cookielib import CookieJar, Cookie
-    from Cookie import SimpleCookie, BaseCookie, CookieError
+    from Cookie import SimpleCookie
 
     string_type = unicode
     ascii_letters = string.letters
@@ -389,9 +389,6 @@ ALL_HEADER_FIELDS = CLIENT_HEADER_FIELDS.union(SERVER_HEADER_FIELDS)
 ALL_HEADER_FIELDS_DICT = dict(((k.lower(), k) for k in ALL_HEADER_FIELDS))
 CRLF = '\r\n'
 LWS = '\r\n '
-HEADER_FIELDS_JOINER = {'Set-Cookie': None,
-                        'Set-Cookie2': None}
-
 TYPE_HEADER_FIELDS = {'client': CLIENT_HEADER_FIELDS,
                       'server': SERVER_HEADER_FIELDS,
                       'both': ALL_HEADER_FIELDS}
@@ -405,6 +402,15 @@ def capheader(name):
 
 
 def header_field(name, HEADERS_SET=None, strict=False):
+    '''Return a header `name` in Camel case.
+
+    For example::
+
+        header_field('connection') == 'Connection'
+        header_field('accept-charset') == 'Accept-Charset'
+
+    If ``header_set`` is given, only return headers included in the set.
+    '''
     name = name.lower()
     if name.startswith('x-'):
         return capheader(name)
@@ -418,13 +424,37 @@ def header_field(name, HEADERS_SET=None, strict=False):
             return capheader(name)
 
 
+####    HEADERS UTILITIES
+HEADER_FIELDS_JOINER = {'Cookie': '; ',
+                        'Set-Cookie': None,
+                        'Set-Cookie2': None}
+
+def split_comma(value):
+    return [v for v in (v.strip() for v in value.split(',')) if v]
+
+def parse_cookies(value):
+    return [c.OutputString() for c in SimpleCookie(value).values()]
+
+header_parsers = {'Connection': split_comma,
+                  'Cookie': parse_cookies}
+
+
+def header_values(header, value):
+    assert isinstance(value, str)
+    if header in header_parsers:
+        return header_parsers[header](value)
+    else:
+        return [value]
+
+
 def quote_header_value(value, extra_chars='', allow_token=True):
     """Quote a header value if necessary.
 
-:param value: the value to quote.
-:param extra_chars: a list of extra characters to skip quoting.
-:param allow_token: if this is enabled token values are returned
-                    unchanged."""
+    :param value: the value to quote.
+    :param extra_chars: a list of extra characters to skip quoting.
+    :param allow_token: if this is enabled token values are returned
+        unchanged.
+    """
     value = force_native_str(value)
     if allow_token:
         token_chars = HEADER_TOKEN_CHARS | set(extra_chars)
@@ -434,10 +464,13 @@ def quote_header_value(value, extra_chars='', allow_token=True):
 
 
 def unquote_header_value(value, is_filename=False):
-    """Unquotes a header value. Reversal of :func:`quote_header_value`.
-This does not use the real unquoting but what browsers are actually
-using for quoting.
-:param value: the header value to unquote."""
+    """Unquotes a header value.
+
+    Reversal of :func:`quote_header_value`. This does not use the real
+    un-quoting but what browsers are actually using for quoting.
+
+    :param value: the header value to unquote.
+    """
     if value and value[0] == value[-1] == '"':
         # this is not the real unquoting, but fixing this so that the
         # RFC is met will result in bugs with internet explorer and
@@ -485,21 +518,6 @@ def parse_dict_header(value):
             value = unquote_header_value(value[1:-1])
         result[name] = value
     return result
-
-
-def split_comma(value):
-    return [v for v in (v.strip() for v in value.split(',')) if v]
-
-
-header_parsers = {'connection': split_comma}
-
-
-def parse_header(header, value):
-    assert isinstance(value, str)
-    if header in header_parsers:
-        return header_parsers[header](value)
-    else:
-        return [value]
 
 
 class Headers(object):
@@ -560,10 +578,6 @@ class Headers(object):
     def __bytes__(self):
         return str(self).encode(DEFAULT_CHARSET)
 
-    def __iter__(self):
-        for k, values in iteritems(self._headers):
-            yield k, ', '.join(values)
-
     def __len__(self):
         return len(self._headers)
 
@@ -572,7 +586,7 @@ class Headers(object):
         return header_type_to_int.get(self.kind)
 
     def update(self, iterable):
-        """Extend the headers.
+        """Extend the headers with an ``iterable``.
 
         :param iterable: a dictionary or an iterable over keys, vaues tuples.
         """
@@ -600,7 +614,12 @@ class Headers(object):
         return header_field(key) in self._headers
 
     def __getitem__(self, key):
-        return ', '.join(self._headers[header_field(key)])
+        key = header_field(key)
+        values = self._headers[key]
+        joiner = HEADER_FIELDS_JOINER.get(key, ', ')
+        if joiner is None:
+            joiner = '; '
+        return joiner.join(values)
 
     def __delitem__(self, key):
         self._headers.__delitem__(header_field(key))
@@ -609,7 +628,7 @@ class Headers(object):
         key = header_field(key, self.all_headers, self.strict)
         if key and value:
             if not isinstance(value, list):
-                value = parse_header(key.lower(), value)
+                value = header_values(key, value)
             self._headers[key] = value
 
     def get(self, key, default=None):
@@ -671,22 +690,22 @@ class Headers(object):
         '''
         return self._headers.get(header_field(key), [])
 
-    def add_header(self, key, value, **params):
-        '''Add ``value`` to ``key`` header.
+    def add_header(self, key, values):
+        '''Add ``values`` to ``key`` header.
 
         If the header is already available, append the value to the list.
+        :param values: a string value or a list/tuple of strings values
+            for header ``key``
         '''
         key = header_field(key, self.all_headers, self.strict)
-        if key and value:
-            assert isinstance(value, str)
-            values = self._headers.get(key, [])
-            value = value.strip()
-            if params:
-                value = '%s; %s' % (value, '; '.join(('%s=%s' % kv for kv
-                                                      in params.items())))
-            if value not in values:
-                values.append(value)
-            self._headers[key] = values
+        if key and values:
+            if not isinstance(values, (tuple, list)):
+                values = header_values(key, values)
+            current = self._headers.get(key, [])
+            for value in values:
+                if value and not value in current:
+                    current.append(value)
+            self._headers[key] = current
 
     def remove_header(self, key, value=None):
         '''Remove the header at ``key``.
@@ -712,6 +731,16 @@ class Headers(object):
         '''Full headers bytes representation'''
         vs = version + (status, self)
         return ('HTTP/%s.%s %s\r\n%s' % vs).encode(DEFAULT_CHARSET)
+
+    def __iter__(self):
+        dj = ', '
+        for k, values in iteritems(self._headers):
+            joiner = HEADER_FIELDS_JOINER.get(k, dj)
+            if joiner:
+                yield k, joiner.join(values)
+            else:
+                for value in values:
+                    yield k, value
 
     def _ordered(self):
         hf = HEADER_FIELDS
@@ -768,30 +797,10 @@ class InvalidChunkSize(Exception):
 class HttpParser(object):
     '''A python HTTP parser.
 
-Original code from https://github.com/benoitc/http-parser
+    Original code from https://github.com/benoitc/http-parser
 
-2011 (c) Benoit Chesneau <benoitc@e-engura.org>
-
-Permission is hereby granted, free of charge, to any person
-obtaining a copy of this software and associated documentation
-files (the "Software"), to deal in the Software without
-restriction, including without limitation the rights to use,
-copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the
-Software is furnished to do so, subject to the following
-conditions:
-
-The above copyright notice and this permission notice shall be
-included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-OTHER DEALINGS IN THE SOFTWARE.'''
+    2011 (c) Benoit Chesneau <benoitc@e-engura.org>
+    '''
     def __init__(self, kind=2, decompress=False):
         self.decompress = decompress
         # errors vars
@@ -1016,20 +1025,22 @@ OTHER DEALINGS IN THE SOFTWARE.'''
             name = name.rstrip(" \t").upper()
             if HEADER_RE.search(name):
                 raise InvalidHeader("invalid header name %s" % name)
-            name, value = name.strip().lower(), [value.lstrip()]
+            name, value = header_field(name.strip()), [value.lstrip()]
             # Consume value continuation lines
             while len(lines) and lines[0].startswith((" ", "\t")):
                 value.append(lines.popleft())
             value = ''.join(value).rstrip()
-            # multiple headers
             if name in self._headers:
-                value = "%s, %s" % (self._headers[name], value)
-            # store new header value
-            self._headers[name] = value
+                self._headers[name].append(value)
+            else:
+                self._headers[name] = [value]
         # detect now if body is sent by chunks.
-        clen = self._headers.get('content-length')
-        te = self._headers.get('transfer-encoding', '').lower()
-        self._chunked = (te == 'chunked')
+        clen = self._headers.get('Content-Length')
+        if 'Transfer-Encoding' in self._headers:
+            te = self._headers['Transfer-Encoding'][0].lower()
+            self._chunked = (te == 'chunked')
+        else:
+            self._chunked = False
         #
         status = self._status_code
         if status and (status == httpclient.NO_CONTENT or
@@ -1039,7 +1050,7 @@ OTHER DEALINGS IN THE SOFTWARE.'''
             clen = 0
         elif clen is not None:
             try:
-                clen = int(clen)
+                clen = int(clen[0])
             except ValueError:
                 clen = None
             else:
@@ -1052,8 +1063,8 @@ OTHER DEALINGS IN THE SOFTWARE.'''
             self._clen_rest = self._clen = clen
         #
         # detect encoding and set decompress object
-        if self.decompress:
-            encoding = self._headers.get('content-encoding')
+        if self.decompress and 'Content-Encoding' in self._headers:
+            encoding = self._headers['Content-Encoding'][0]
             if encoding == "gzip":
                 self.__decompress_obj = zlib.decompressobj(16+zlib.MAX_WBITS)
             elif encoding == "deflate":
@@ -1279,28 +1290,6 @@ def cookiejar_from_dict(cookie_dict, cookiejar=None):
         return cookiejar
     else:
         return cookie_dict
-
-
-def parse_cookie(cookie):
-    '''Parse an `HTTP cookie`_ string.
-
-    Return a dictionary of cookie name/values.
-    '''
-    if not cookie:
-        return {}
-    if not isinstance(cookie, BaseCookie):
-        try:
-            c = SimpleCookie()
-            c.load(cookie)
-        except CookieError:     # pragma    nocover
-            # Invalid cookie
-            return {}
-    else:
-        c = cookie
-    cookiedict = {}
-    for key in c.keys():
-        cookiedict[key] = c.get(key).value
-    return cookiedict
 
 
 cc_delim_re = re.compile(r'\s*,\s*')