+++ tests/http/tools.py
@@ -2,7 +2,7 @@
 import time
 
 from pulsar.apps.test import unittest
-from pulsar.utils.httpurl import (Headers, parse_cookie, CacheControl,
+from pulsar.utils.httpurl import (Headers, CacheControl,
                                   urlquote, unquote_unreserved, requote_uri,
                                   remove_double_slash, appendslash, capfirst,
                                   encode_multipart_formdata, http_date,
@@ -133,15 +133,3 @@ class TestTools(unittest.TestCase):
         j = cookiejar_from_dict({'bla': 'foo'})
         j2 = cookiejar_from_dict({'pippo': 'pluto'}, j)
         self.assertEqual(j, j2)
-
-    def test_parse_cookie(self):
-        self.assertEqual(parse_cookie('invalid key=true'),
-                         {'key':'true'})
-        self.assertEqual(parse_cookie('invalid;key=true'),
-                         {'key':'true'})
-        self.assertEqual(parse_cookie(''), {})
-        self.assertEqual(parse_cookie(None), {})
-        c = SimpleCookie()
-        c.load('key=true')
-        self.assertEqual(parse_cookie(c), {'key':'true'})
-        self.assertEqual(parse_cookie('key='), {'key': ''})